clicking the delete button will show an alert

+++ src/pages/MyToys/MyToys.jsx
@@ -1,4 +1,5 @@
 import React, { useContext, useEffect, useState } from 'react';
+import Swal from 'sweetalert2';
 import { AuthContext } from '../../providers/AuthProvider';
 import MyToysTable from '../MyToysTable/MyToys';
 
@@ -16,15 +17,33 @@ const MyToys = () => {
     }, [user])
 
     const handleDelete = id => {
-        fetch(`http://localhost:5000/toyRemove/${id}`, {
-            method: 'DELETE'
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/toyRemove/${id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount > 0) {
+                            Swal.fire(
+                                'Deleted!',
+                                'Your file has been deleted.',
+                                'success'
+                            )
+                        }
+                        const check = myToys.filter(toy => toy._id !== id);
+                        setMyToys(check)
+                    })
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data)
-                const check = myToys.filter(toy => toy._id !== id);
-                setMyToys(check)
-            })
     }
 
     return (
